requirements.txt
absl-py==1.3.0
alembic==1.9.1
attrs==22.2.0
autopage==0.5.1
bottle==0.12.23
Box2D==2.3.10
cliff==4.1.0
cloudpickle==2.2.0
cmaes==0.9.1
cmd2==2.4.2
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
greenlet==2.0.1
gym==0.21.0
importlib-metadata==4.13.0
importlib-resources==5.10.2
joblib==1.2.0
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
optuna==3.0.5
optuna-dashboard==0.8.1
packaging==22.0
pandas==1.5.2
patsy==0.5.3
pbr==5.11.0
Pillow==9.3.0
plotly==5.11.0
prettytable==3.6.0
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
rich==13.0.0
scikit-learn==1.2.0
scipy==1.8.1
simpy==4.0.1
six==1.16.0
SQLAlchemy==1.4.46
stable-baselines3==1.6.2
statsmodels==0.13.5
stevedore==4.1.1
tenacity==8.1.0
threadpoolctl==3.1.0
torch==1.13.0
tqdm==4.64.1
typing_extensions==4.4.0
wcwidth==0.2.5
zipp==3.11.0
